requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
fastapi
uvicorn[standard]
torch
torchvision
pillow
pulp
python-multipart
aiofiles
scikit-image
